+++ src/unused-exports/settings.ts
@@ -1,0 +1,5 @@
+import * as vscode from 'vscode';
+
+export function isResultExpanded(): boolean {
+  return vscode.workspace.getConfiguration().get('findUnusedExports.defaultResultExpanded', false);
+}